Fix merge issues

+++ src/libs/dialog/include/dialog.h
@@ -65,7 +65,7 @@ class DIALOG final : public Entity
 
     static void AddToStringArrayLimitedByWidth(const std::string_view &text, int32_t nFontID, float fScale, int32_t nLimitWidth,
                                                std::vector<std::string> &asOutTextList, VDX9RENDER *render_service,
-                                               std::vector<long> *panPageIndices, size_t nPageSize);
+                                               std::vector<int32_t> *panPageIndices, size_t nPageSize);
 
   private:
     void EmergencyExit();
